Initial Commit
NonWorking State: The Scene3D element is not rendering I presume due to missing libraries on my machine: QtDataVisulization, specifically.
I'm not certain on how easy this will be to add due to how github.com/therecipe/qt has it's API's built; it does however work, enough to make this easier for me to grasp.
Long term concern is the licensing for the source code, If I were to try and publish something as a business I may have to reimplement many of the underlying code. We will cross that bridge when we get there.